Add api permissions (#518)
* Add permissions to the API

* Optimize imports

* Optimize more imports

* Update authorizer and tests

* Refactor code and test

* CreateSubjectApiTest now passes

* Remove unnecessary code and update tests

* Fix if statement

* Add SubjectActionAuthorizerInterface, refactor code and tests

* Rename SubjectActionAuthorizerInterface

* Rename authorizer

+++ Neo/src/Application/Actions/DeleteSubject/DeleteSubjectAction.php
@@ -6,15 +6,20 @@ namespace ProfessionalWiki\NeoWiki\Application\Actions\DeleteSubject;
 
 use ProfessionalWiki\NeoWiki\Application\SubjectRepository;
 use ProfessionalWiki\NeoWiki\Domain\Subject\SubjectId;
+use ProfessionalWiki\NeoWiki\Infrastructure\SubjectActionAuthorizer;
 
 class DeleteSubjectAction {
-
 	public function __construct(
-		private SubjectRepository $subjectRepository
+		private SubjectRepository $subjectRepository,
+		private SubjectActionAuthorizer $subjectActionAuthorizer
 	) {
 	}
 
 	public function deleteSubject( SubjectId $subjectId ): void {
+		if ( !$this->subjectActionAuthorizer->canDeleteSubject() ) {
+			throw new \RuntimeException( 'You do not have the necessary permissions to delete this subject' );
+		}
+
 		$this->subjectRepository->deleteSubject( $subjectId );
 	}
